backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.6
pillow==10.2.0
numpy==1.26.3
pytesseract==0.3.10
easyocr==1.7.1
transformers==4.38.1
torch==2.1.2
torchvision==0.16.2
accelerate==0.26.1
opencv-python==4.9.0.80
pydantic==2.6.1
python-sane==2.9.1
